Configurable timings, fit vert scrollbar
It was choosing sizes that needed a horizontal scrollbar anytime I had
enough files visible to need a vertical scrollbar. I've attempted to
avoid this by including the vertical scrollbar width in the default
padding. Only the default tree is changed because I don't use Nuclide.

Also create some settings to allow the user to speed up or slow down
the animations if desired, defaulting to the previously hard coded
values.